+++ spec/spec_helper.rb
@@ -1,1 +1,8 @@
+# frozen_string_literal: true
+
 require_relative "../lib/lumberjack_json_device"
+
+RSpec.configure do |config|
+  config.warnings = true
+  config.order = :random
+end